Qual: Renamed french field "envente" into "tosell".

+++ htdocs/install/mysql/migration/2.9.0-3.0.0.sql
@@ -12,6 +12,6 @@ alter table llx_product       add column recuperableonly integer NOT NULL DEFAUL
 alter table llx_product_price add column recuperableonly integer NOT NULL DEFAULT '0' after tva_tx;
 
  
-alter table llx_product change column envente tosell tinyint DEFAULT 1;
-alter table llx_product add column tobuy tinyint DEFAULT 1 after tosell;
+alter table llx_product change column envente tosell smallint DEFAULT 1;
+alter table llx_product add column tobuy smallint DEFAULT 1 after tosell;
